# perf: Batch Plain Parquet UTF-8 verification