```
$ POLARS_MAX_THREADS=1 poop './plparbench-before 10 plain.parquet auto' './plparbench-after 10 plain.parquet auto'

Benchmark 1 (3 runs): ./plparbench-before 10 plain.parquet auto
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          13.0s  ±  701ms    12.5s  … 13.8s           0 ( 0%)        0%
  peak_rss           1.46GB ±  407KB    1.46GB … 1.46GB          0 ( 0%)        0%
  cpu_cycles         37.9G  ± 84.2M     37.8G  … 38.0G           0 ( 0%)        0%
  instructions        113G  ± 46.7M      113G  …  113G           0 ( 0%)        0%
  cache_references    426M  ±  397K      425M  …  426M           0 ( 0%)        0%
  cache_misses        111M  ±  777K      110M  …  112M           0 ( 0%)        0%
  branch_misses       634K  ± 53.6K      597K  …  695K           0 ( 0%)        0%
Benchmark 2 (3 runs): ./plparbench-after 10 plain.parquet auto
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          8.55s  ± 36.1ms    8.51s  … 8.58s           0 ( 0%)        ⚡- 34.2% ±  8.7%
  peak_rss           1.46GB ±  179KB    1.46GB … 1.46GB          0 ( 0%)          -  0.0% ±  0.0%
  cpu_cycles         20.6G  ± 3.93M     20.6G  … 20.6G           0 ( 0%)        ⚡- 45.6% ±  0.4%
  instructions       50.0G  ± 11.1M     50.0G  … 50.0G           0 ( 0%)        ⚡- 55.6% ±  0.1%
  cache_references    377M  ±  299K      377M  …  378M           0 ( 0%)        ⚡- 11.4% ±  0.2%
  cache_misses        102M  ± 94.0K      102M  …  102M           0 ( 0%)        ⚡-  8.1% ±  1.1%
  branch_misses       237K  ± 8.80K      229K  …  247K           0 ( 0%)        ⚡
```

For `PLAIN` string page decoding, this can significantly improve the performance.